+++ ChangeLog.md
@@ -5,6 +5,12 @@
 
 # CaPyCli - Clearing Automation Python Command Line Tool for SW360
 
+## NEXT
+
+* `bom map`: The options `--dbx` and `-all` were replaced by `--matchmode`.
+* `bom map`: new `--matchmode` options `full-search` (report all best matches) and
+  `qualifier-match` (consider PackageURL qualifiers). See `Readme_Mapping.md`.
+
 ## 2.9.1
 
 * `bom map` will provide the `purl` from SW360 in the output BOM's components